Reorganizing sources into multiple files
