+++ Sources/MissionControl.swift
@@ -24,12 +24,10 @@
 
 import Foundation
 
-// MARK: - MissionControl
-
 /// Facade class for using MissionControl.
 public class MissionControl {
     
-    // MARK: Types
+    // MARK: - Types
     
     /// Errors types which can be throwed when refreshing local config from remote.
     public enum ServerError: Error {
@@ -57,7 +55,7 @@ public class MissionControl {
         }
     }
     
-    // MARK: Properties
+    // MARK: - Properties
     
     /// Delegate for Mission Control.
     public class var delegate: MissionControlDelegate? {
@@ -85,7 +83,7 @@ public class MissionControl {
         return ACMissionControl.shared.cacheDate
     }
     
-    // MARK: API
+    // MARK: - API
     
     /**
         This should be called on your app start to initialize and/or refresh remote config.
@@ -113,8 +111,6 @@ public class MissionControl {
     
 }
 
-// MARK: - MissionControlDelegate
-
 /**
     Delegate for Mission Control.
  
@@ -135,331 +131,4 @@ public protocol MissionControlDelegate: class {
         - parameter error: Error which happened during config refresh from remote.
     */
     func missionControlDidFailRefreshingConfig(error: Error)
-}
-
-// MARK: - Custom Types
-
-/// Block which throws via inner block.
-public typealias ThrowWithInnerBlock = (() throws -> Void) -> Void
-
-/// Block which throws dictionary via inner block.
-public typealias ThrowJSONWithInnerBlock = (_ block: @escaping () throws -> [String : AnyObject]) -> Void
-
-// MARK: - Accessors
-
-/**
-    Accessor for retreiving setting of generic type `T` for given key.
-
-    This method will resolve to proper setting by following this priority order:
-    1. Remote setting from memory (received in the last refresh).
-    2. Remote setting from disk cache (if never refreshed in current app session (ex. offline)).
-    3. Local setting from disk (defaults provided in `localConfig` on MissionControl `launch`).
-    4. Provided fallback value (if provided)
-
-    - parameter key: Key for the setting.
-    - parameter fallback: Fallback value if setting is not available in any config.
-
-    - returns: Resolved setting of generic type `T` for given key.
-*/
-public func ConfigGeneric<T>(_ key: String, fallback: T) -> T {
-    if let remoteValue = ACMissionControl.shared.remoteConfig?[key] as? T {
-        return remoteValue
-    } else if let cachedValue = ACMissionControl.shared.cachedConfig?[key] as? T {
-        return cachedValue
-    } else if let localValue = ACMissionControl.shared.localConfig?[key] as? T {
-        return localValue
-    } else {
-        return fallback
-    }
-}
-
-/**
-    Async "Force Remote" Accessor for retreiving the latest setting of generic type `T` for given key.
- 
-    This method will first call `refresh` method after which it will evaluate its success.
- 
-    If `refresh` was successful, it will call normal accessor of generic type `T` for given key,
-    which will by its priority order resolve to the latest remote value as a parameter inside `completion` handler.
- 
-    If `refresh` fails, it will return provided `fallback` value as a parameter inside `completion` block.
-
-    - parameter key: Key for the setting.
-    - parameter fallback: Fallback value of generic type `T` if refresh is not successful.
-*/
-public func ConfigGenericForce<T>(_ key: String, fallback: T, completion: @escaping ((_ forced: T) -> Void)) {
-    MissionControl.refresh({ (innerBlock) in
-        do {
-            let _ = try innerBlock()
-            completion(ConfigGeneric(key, fallback: fallback))
-        } catch {
-            completion(fallback)
-        }
-    })
-}
-
-/**
-    Accessor helper for retreiving setting of type `Bool` for given key.
-    It will call `ConfigGeneric<T>` with `Bool` type.
-
-    - parameter key: Key for the setting.
-    - parameter fallback: Fallback value if setting not available in any config. Defaults to `Bool()`.
-
-    - returns: Resolved setting of type `Bool` for given key.
-*/
-public func ConfigBool(_ key: String, fallback: Bool = Bool()) -> Bool {
-    return ConfigGeneric(key, fallback: fallback)
-}
-
-/**
-    Async "Force Remote" Accessor helper for retreiving the latest setting of type `Bool` for given key.
-    It will call `ConfigGenericForce<T>` with `Bool` type.
- 
-    - parameter key: Key for the setting.
-    - parameter fallback: Fallback value if refresh was not successful.
-*/
-public func ConfigBoolForce(_ key: String, fallback: Bool, completion: @escaping ((_ forced: Bool) -> Void)) {
-    ConfigGenericForce(key, fallback: fallback, completion: completion)
-}
-
-/**
-    Accessor helper for retreiving setting of type `Int` for given key.
-    It will call `ConfigGeneric<T>` with `Int` type.
-
-    - parameter key: Key for the setting.
-    - parameter fallback: Fallback value if setting not available in any config. Defaults to `Int()`.
-
-    - returns: Resolved setting of type `Int` for given key.
-*/
-public func ConfigInt(_ key: String, fallback: Int = Int()) -> Int {
-    return ConfigGeneric(key, fallback: fallback)
-}
-
-/**
-    Async "Force Remote" Accessor helper for retreiving the latest setting of type `Int` for given key.
-    It will call `ConfigGenericForce<T>` with `Int` type.
-
-    - parameter key: Key for the setting.
-    - parameter fallback: Fallback value if refresh was not successful.
-*/
-public func ConfigIntForce(_ key: String, fallback: Int, completion: @escaping ((_ forced: Int) -> Void)) {
-    ConfigGenericForce(key, fallback: fallback, completion: completion)
-}
-
-/**
-    Accessor helper for retreiving setting of type `Double` for given key.
-    It will call `ConfigGeneric<T>` with `Double` type.
-
-    - parameter key: Key for the setting.
-    - parameter fallback: Fallback value if setting not available in any config. Defaults to `Double()`.
-
-    - returns: Resolved setting of type `Double` for given key.
-*/
-public func ConfigDouble(_ key: String, fallback: Double = Double()) -> Double {
-    return ConfigGeneric(key, fallback: fallback)
-}
-
-/**
-    Async "Force Remote" Accessor helper for retreiving the latest setting of type `Double` for given key.
-    It will call `ConfigGenericForce<T>` with `Double` type.
-
-    - parameter key: Key for the setting.
-    - parameter fallback: Fallback value if refresh was not successful.
-*/
-public func ConfigDoubleForce(_ key: String, fallback: Double, completion: @escaping ((_ forced: Double) -> Void)) {
-    ConfigGenericForce(key, fallback: fallback, completion: completion)
-}
-
-/**
-    Accessor helper for retreiving setting of type `String` for given key.
-    It will call `ConfigGeneric<T>` with `String` type.
-
-    - parameter key: Key for the setting.
-    - parameter fallback: Fallback value if setting not available in any config. Defaults to `String()`.
-
-    - returns: Resolved setting of type `String` for given key.
-*/
-public func ConfigString(_ key: String, fallback: String = String()) -> String {
-    return ConfigGeneric(key, fallback: fallback)
-}
-
-/**
-    Async "Force Remote" Accessor helper for retreiving the latest setting of type `String` for given key.
-    It will call `ConfigGenericForce<T>` with `String` type.
-
-    - parameter key: Key for the setting.
-    - parameter fallback: Fallback value if refresh was not successful.
-*/
-public func ConfigStringForce(_ key: String, fallback: String, completion: @escaping ((_ forced: String) -> Void)) {
-    ConfigGenericForce(key, fallback: fallback, completion: completion)
-}
-
-// MARK: - ACMissionControl
-
-class ACMissionControl {
-    
-    // MARK: Singleton
-    
-    static let shared = ACMissionControl()
-    
-    // MARK: Properties
-    
-    weak var delegate: MissionControlDelegate?
-    
-    var localConfig: [String : Any]?
-    
-    var remoteURL: URL? {
-        didSet {
-            if let _ = remoteURL {
-                refresh({ (block) in
-                    do {
-                        _ = try block()
-                    } catch {
-                        print(error)
-                    }
-                })
-            }
-        }
-    }
-    
-    var remoteConfig: [String : Any]? {
-        didSet {
-            if let newConfig = remoteConfig {
-                refreshDate = Date()
-                
-                cachedConfig = newConfig
-                cacheDate = refreshDate
-
-                informListeners(oldConfig: oldValue, newConfig: newConfig)
-            }
-        }
-    }
-    
-    private func informListeners(oldConfig: [String : Any]?, newConfig: [String : Any]) {
-        let userInfo = userInfoWithConfig(old: oldConfig, new: newConfig)
-        delegate?.missionControlDidRefreshConfig(old: oldConfig, new: newConfig)
-        sendNotification(MissionControl.Notification.DidRefreshConfig, userInfo: userInfo)
-    }
-    
-    var refreshDate: Date?
-    
-    private struct Cache {
-        static let Config = "ACMissionControl.CachedConfig"
-        static let Date = "ACMissionControl.CacheDate"
-    }
-    
-    var cachedConfig: [String : Any]? {
-        get {
-            let userDefaults = UserDefaults.standard
-            let config = userDefaults.object(forKey: Cache.Config) as? [String : AnyObject]
-            return config
-        }
-        set {
-            let userDefaults = UserDefaults.standard
-            userDefaults.set(newValue, forKey: Cache.Config)
-            userDefaults.synchronize()
-        }
-    }
-    
-    var cacheDate: Date? {
-        get {
-            let userDefaults = UserDefaults.standard
-            let config = userDefaults.object(forKey: Cache.Date) as? Date
-            return config
-        }
-        set {
-            let userDefaults = UserDefaults.standard
-            userDefaults.set(newValue, forKey: Cache.Date)
-            userDefaults.synchronize()
-        }
-    }
-    
-    // MARK: API
-    
-    func refresh(_ completion: ThrowWithInnerBlock? = nil) {
-        getRemoteConfig { [unowned self] (block) in
-            DispatchQueue.main.async { [unowned self] in
-                do {
-                    let remoteConfig = try block()
-                    self.remoteConfig = remoteConfig
-                    completion?({ })
-                } catch {
-                    self.informListeners(error)
-                    completion?({ throw error })
-                }
-            }
-        }
-    }
-    
-    private func informListeners(_ error: Error) {
-        delegate?.missionControlDidFailRefreshingConfig(error: error)
-        let userInfo: [AnyHashable : Any] = ["Error" : "\(error)"]
-        sendNotification(MissionControl.Notification.DidFailRefreshingConfig, userInfo: userInfo)
-    }
-    
-    // MARK: Helpers
-    
-    func resetAll() {
-        localConfig = nil
-        cachedConfig = nil
-        remoteConfig = nil
-        refreshDate = nil
-        remoteURL = nil
-        delegate = nil
-    }
-    
-    func resetRemote() {
-        remoteConfig = nil
-        refreshDate = nil
-    }
-    
-    private func userInfoWithConfig(old: [AnyHashable : Any]?, new: [AnyHashable : Any]?) -> [AnyHashable : Any]? {
-        if old == nil && new == nil {
-            return nil
-        } else {
-            var userInfo = [AnyHashable : Any]()
-            if let oldConfig = old {
-                userInfo[MissionControl.Notification.UserInfo.OldConfigKey] = oldConfig
-            }
-            if let newConfig = new {
-                userInfo[MissionControl.Notification.UserInfo.NewConfigKey] = newConfig
-            }
-            return userInfo
-        }
-    }
-    
-    private func sendNotification(_ name: String, userInfo: [AnyHashable : Any]? = nil) {
-        let center = NotificationCenter.default
-        center.post(name: Notification.Name(rawValue: name), object: self, userInfo: userInfo)
-    }
-    
-    private func getRemoteConfig(_ completion: @escaping ThrowJSONWithInnerBlock) {
-        guard let url = remoteURL
-            else { completion({ throw MissionControl.ServerError.noRemoteURL }); return }
-    
-        let request = URLRequest(url: url)
-        let session = URLSession.shared
-        
-        let task = session.dataTask(with: request) { [unowned self] (data, response, error) in
-            guard let httpResponse = response as? HTTPURLResponse, httpResponse.statusCode == 200
-            else { completion({ throw MissionControl.ServerError.badResponseCode }); return }
-            self.parseRemoteConfigFromData(data, completion: completion)
-        }
-        
-        task.resume()
-    }
-    
-    private func parseRemoteConfigFromData(_ data: Data?, completion: ThrowJSONWithInnerBlock) {
-        guard let configData = data
-            else { completion({ throw MissionControl.ServerError.invalidData }); return }
-        
-        do {
-            let json = try JSONSerialization.jsonObject(with: configData, options: .allowFragments)
-            guard let config = json as? [String : AnyObject]
-                else { completion({ throw MissionControl.ServerError.invalidData }); return }
-            completion({ return config })
-        } catch {
-            completion({ throw MissionControl.ServerError.invalidData })
-        }
-    }
-    
 }
